End of secttion 18

+++ src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -17,7 +18,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor( private messageService: MessageService ) { }
+  constructor( private messageService: MessageService, private confirmService: ConfirmService ) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -36,9 +37,16 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage( id: number ) {
-    this.messageService.deleteMessage( id )
+    this.confirmService.confirm( 'Confirm delete message', 'This cannot be undone' )
       .subscribe(
-        () => this.messages.splice( this.messages.findIndex( m => m.id === id ), 1 )
+        result => {
+          if ( result ) {
+            this.messageService.deleteMessage( id )
+              .subscribe(
+                () => this.messages.splice( this.messages.findIndex( m => m.id === id ), 1 )
+              );
+          }
+        }
       );
   }
 
@@ -46,5 +54,5 @@ export class MessagesComponent implements OnInit {
     this.pageNumber = event.page;
     this.loadMessages();
   }
-  
+
 }
